cleanup/unify push method in views update lints move settings widgets into separate files
