+++ lib/views/unconfirmed_transaction.dart
@@ -1,21 +1,28 @@
+import 'dart:async';
+
+import 'package:cupcake/coins/abstract/address.dart';
+import 'package:cupcake/coins/abstract/amount.dart';
+import 'package:cupcake/coins/abstract/wallet.dart';
 import 'package:cupcake/utils/call_throwable.dart';
 import 'package:cupcake/view_model/unconfirmed_transaction_view_model.dart';
 import 'package:cupcake/views/abstract.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-//ignore: must_be_immutable
 class UnconfirmedTransactionView extends AbstractView {
-  static Future<void> staticPush(
-      BuildContext context, UnconfirmedTransactionViewModel viewModel) async {
-    await Navigator.of(context).push(
-      CupertinoPageRoute(
-        builder: (context) => UnconfirmedTransactionView(viewModel: viewModel),
-      ),
-    );
-  }
-
-  UnconfirmedTransactionView({super.key, required this.viewModel});
+  UnconfirmedTransactionView({
+    super.key,
+    required CoinWallet wallet,
+    required Amount fee,
+    required Map<Address, Amount> destMap,
+    required FutureOr<void> Function(BuildContext) confirmCallback,
+    required FutureOr<void> Function(BuildContext) cancelCallback,
+  }) : viewModel = UnconfirmedTransactionViewModel(
+          wallet: wallet,
+          fee: fee,
+          destMap: destMap,
+          confirmCallback: confirmCallback,
+          cancelCallback: cancelCallback,
+        );
 
   @override
   final UnconfirmedTransactionViewModel viewModel;
