+++ lib/views/widgets/settings/boolean_config_element.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class BooleanConfigElement extends StatelessWidget {
+  const BooleanConfigElement({
+    super.key,
+    required this.title,
+    required this.subtitleEnabled,
+    required this.subtitleDisabled,
+    required this.value,
+    required this.onChange,
+  });
+
+  final String title;
+  final String subtitleEnabled;
+  final String subtitleDisabled;
+  final bool value;
+  final Function(bool val) onChange;
+  @override
+  Widget build(BuildContext context) {
+    return CheckboxListTile(
+      title: Text(title),
+      subtitle: Text(value ? subtitleEnabled : subtitleDisabled),
+      value: value,
+      onChanged: (bool? value) {
+        onChange(value == true);
+      },
+    );
+  }
+}
