+++ lib/views/widgets/settings/integer_config_element.dart
@@ -1,0 +1,46 @@
+import 'package:cupcake/utils/alerts/widget.dart';
+import 'package:cupcake/utils/config.dart';
+import 'package:flutter/material.dart';
+
+class IntegerConfigElement extends StatelessWidget {
+  IntegerConfigElement({
+    super.key,
+    required this.title,
+    required this.hint,
+    required this.value,
+    required this.onChange,
+  });
+
+  final String title;
+  final String? hint;
+  final int value;
+  final Function(int val) onChange;
+  late final ctrl = TextEditingController(text: value.toString());
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: Text(title),
+      onLongPress: () {
+        CupcakeConfig.instance.debug = true;
+        CupcakeConfig.instance.save();
+      },
+      subtitle: TextField(
+        controller: ctrl,
+        onSubmitted: (String value) {
+          final i = int.tryParse(value);
+          if (i == null) return;
+          onChange(i);
+        },
+      ),
+      trailing: hint == null
+          ? null
+          : IconButton(
+              icon: const Icon(Icons.info),
+              onPressed: () {
+                showAlertWidget(
+                    context: context, title: title, body: [Text(hint ?? "")]);
+              },
+            ),
+    );
+  }
+}
